release v0.24.2

+++ src/io/uring/Queue.cxx
@@ -1,34 +1,6 @@
-/*
- * Copyright 2020 CM4all GmbH
- * All rights reserved.
- *
- * author: Max Kellermann <mk@cm4all.com>
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- * - Redistributions of source code must retain the above copyright
- * notice, this list of conditions and the following disclaimer.
- *
- * - Redistributions in binary form must reproduce the above copyright
- * notice, this list of conditions and the following disclaimer in the
- * documentation and/or other materials provided with the
- * distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * ``AS IS'' AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
- * FOR A PARTICULAR PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL THE
- * FOUNDATION OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
- * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
- * STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
- * OF THE POSSIBILITY OF SUCH DAMAGE.
- */
+// SPDX-License-Identifier: BSD-2-Clause
+// Copyright CM4all GmbH
+// author: Max Kellermann <mk@cm4all.com>
 
 #include "Queue.hxx"
 #include "CancellableOperation.hxx"
@@ -40,6 +12,11 @@ namespace Uring {
 
 Queue::Queue(unsigned entries, unsigned flags)
 	:ring(entries, flags)
+{
+}
+
+Queue::Queue(unsigned entries, struct io_uring_params &params)
+	:ring(entries, params)
 {
 }
 
@@ -55,7 +32,7 @@ Queue::RequireSubmitEntry()
 	if (sqe == nullptr) {
 		/* the submit queue is full; submit it to the kernel
 		   and try again */
-		Submit();
+		ring.Submit();
 
 		sqe = GetSubmitEntry();
 		if (sqe == nullptr)
@@ -74,17 +51,25 @@ Queue::AddPending(struct io_uring_sqe &sqe,
 	io_uring_sqe_set_data(&sqe, c);
 }
 
-void
-Queue::DispatchOneCompletion(struct io_uring_cqe &cqe) noexcept
+inline void
+Queue::_DispatchOneCompletion(const struct io_uring_cqe &cqe) noexcept
 {
 	void *data = io_uring_cqe_get_data(&cqe);
 	if (data != nullptr) {
 		auto *c = (CancellableOperation *)data;
-		c->OnUringCompletion(cqe.res);
-		c->unlink();
-		delete c;
+		const bool more = cqe.flags & IORING_CQE_F_MORE;
+		c->OnUringCompletion(cqe.res, more);
+		if (!more) {
+			c->unlink();
+			delete c;
+		}
 	}
+}
 
+void
+Queue::DispatchOneCompletion(struct io_uring_cqe &cqe) noexcept
+{
+	_DispatchOneCompletion(cqe);
 	ring.SeenCompletion(cqe);
 }
 
@@ -99,6 +84,14 @@ Queue::DispatchOneCompletion()
 	return true;
 }
 
+inline unsigned
+Queue::DispatchCompletions(struct io_uring_cqe &_cqe) noexcept
+{
+	return ring.VisitCompletions(&_cqe, [](const struct io_uring_cqe &cqe){
+		_DispatchOneCompletion(cqe);
+	});
+}
+
 bool
 Queue::WaitDispatchOneCompletion()
 {
@@ -108,6 +101,16 @@ Queue::WaitDispatchOneCompletion()
 
 	DispatchOneCompletion(*cqe);
 	return true;
+}
+
+bool
+Queue::SubmitAndWaitDispatchCompletions(struct __kernel_timespec *timeout)
+{
+	auto *cqe = ring.SubmitAndWaitCompletion(timeout);
+	if (cqe == nullptr)
+		return false;
+
+	return DispatchCompletions(*cqe) > 0;
 }
 
 } // namespace Uring
